pyproject.toml
[tool.poetry]
name = "lighthouse-api"
version = "0.1.0"
description = ""
authors = ["Jens Tinglev <tinglev@kth.se>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
azure-storage-blob = "^12.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
